Move currently-unused email invite form to own component
