+++ src/routes/_user/friends.invite.tsx
@@ -1,12 +1,7 @@
 import { useState } from 'react'
 import { createFileRoute, Link } from '@tanstack/react-router'
-import { useMutation } from '@tanstack/react-query'
-import { z } from 'zod'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { Controller, useForm } from 'react-hook-form'
-import toast from 'react-hot-toast'
 
-import { Mail, Phone, Search, Send, Share } from 'lucide-react'
+import { Mail, Phone, Search, Share } from 'lucide-react'
 import { Button, buttonVariants } from '@/components/ui/button'
 import {
 	Card,
@@ -15,10 +10,7 @@ import {
 	CardHeader,
 	CardTitle,
 } from '@/components/ui/card'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
 import { ShowError } from '@/components/errors'
-import supabase from '@/lib/supabase-client'
 import { useProfile } from '@/lib/use-profile'
 
 export const Route = createFileRoute('/_user/friends/invite')({
@@ -60,10 +52,11 @@ function InviteFriendPage() {
 
 function ShareButtons() {
 	const canShare = typeof navigator?.share === 'function'
-	const [error, setError] = useState<null | DOMException | TypeError>(null)
+	const [error, setError] = useState<
+		null | DOMException | TypeError | { message: string }
+	>(null)
 	const { data: profile } = useProfile()
 	const shareData = {
-		// url: 'https://sunlo.app/signup',
 		text: `Hello friend, I'm learning a language with Sunlo, a social language learning app. Will you join me? https://sunlo.app/signup`,
 		title: `Invitation! ${profile?.username} on Sunlo.app`,
 	}
@@ -80,7 +73,12 @@ function ShareButtons() {
 					)
 					// setError(error)
 				})
-		else console.log(`not sharing bc not possible`)
+		else {
+			console.log(`not sharing bc not supported`)
+			setError({
+				message: `Sorry, we can't open your device's share interface. Please try one of the other options.`,
+			})
+		}
 	}
 	return (
 		<div>
@@ -105,67 +103,5 @@ function ShareButtons() {
 			</div>
 			<ShowError>{error?.message}</ShowError>
 		</div>
-	)
-}
-
-const inviteFriendSchema = z.object({
-	email: z.string().email('Please enter a valid email'),
-})
-function InviteFriendForm() {
-	const { control, handleSubmit } = useForm<z.infer<typeof inviteFriendSchema>>(
-		{
-			resolver: zodResolver(inviteFriendSchema),
-		}
-	)
-	// const queryClient = useQueryClient()
-
-	const invite = useMutation({
-		mutationKey: ['user', 'invite_friend'],
-		mutationFn: async (values: z.infer<typeof inviteFriendSchema>) => {
-			const { data, error } = await supabase.auth.admin.inviteUserByEmail(
-				values.email
-			)
-			if (error) throw error
-			return data
-		},
-		onSuccess: (_, values) => {
-			toast.success(`Invitation sent to ${values.email}.`)
-			/*void queryClient.invalidateQueries({
-				queryKey: ['user', 'friend_invited'],
-			})*/
-		},
-	})
-
-	const onSubmit = handleSubmit((data) => {
-		invite.mutate(data)
-	})
-
-	return (
-		<form onSubmit={onSubmit}>
-			<fieldset
-				className="flex flex-row gap-2 items-end"
-				disabled={invite.isPending}
-			>
-				<div className="w-full">
-					<Label htmlFor="email">Friend's email</Label>
-					<Controller
-						name="email"
-						control={control}
-						render={({ field }) => (
-							<Input
-								{...field}
-								type="email"
-								placeholder="Enter your friend's email"
-							/>
-						)}
-					/>
-				</div>
-				<Button disabled={invite.isPending}>
-					<Send />
-					<span className="hidden @md:block">Send</span>
-				</Button>
-			</fieldset>
-			<ShowError className="mt-4">{invite.error?.message}</ShowError>
-		</form>
 	)
 }
